requirements.txt
requests
rsa
debugpy
Pillow
paddleocr
paddlepaddle
numpy
PyQt5
nuitka


